# Add Tracy integration and fix performance regression in MachO x86_64