--- a/src/MachO/DwarfInfo.zig
+++ b/src/MachO/DwarfInfo.zig
@@ -381,71 +381,71 @@ fn findFormSize(self: DwarfInfo, form: u64, di_off: usize, cuh: CompileUnit.Head
     var creader = std.io.countingReader(stream.reader());
     const reader = creader.reader();
 
-    return switch (form) {
+    switch (form) {
         dwarf.FORM.strp,
         dwarf.FORM.sec_offset,
         dwarf.FORM.ref_addr,
-        => if (cuh.is_64bit) @sizeOf(u64) else @sizeOf(u32),
+        => return if (cuh.is_64bit) @sizeOf(u64) else @sizeOf(u32),
 
-        dwarf.FORM.addr => cuh.address_size,
-        dwarf.FORM.exprloc => blk: {
+        dwarf.FORM.addr => return cuh.address_size,
+        dwarf.FORM.exprloc => {
             const expr_len = try leb.readULEB128(u64, reader);
             var i: u64 = 0;
             while (i < expr_len) : (i += 1) {
                 _ = try reader.readByte();
             }
-            break :blk creader.bytes_read;
+            return creader.bytes_read;
         },
-        dwarf.FORM.flag_present => 0,
+        dwarf.FORM.flag_present => return 0,
 
         dwarf.FORM.data1,
         dwarf.FORM.ref1,
         dwarf.FORM.block1,
         dwarf.FORM.flag,
-        => @sizeOf(u8),
+        => return @sizeOf(u8),
 
         dwarf.FORM.data2,
         dwarf.FORM.ref2,
         dwarf.FORM.block2,
-        => @sizeOf(u16),
+        => return @sizeOf(u16),
 
         dwarf.FORM.data4,
         dwarf.FORM.ref4,
         dwarf.FORM.block4,
-        => @sizeOf(u32),
+        => return @sizeOf(u32),
 
         dwarf.FORM.data8,
         dwarf.FORM.ref8,
         dwarf.FORM.ref_sig8,
-        => @sizeOf(u64),
+        => return @sizeOf(u64),
 
         dwarf.FORM.udata,
         dwarf.FORM.ref_udata,
         dwarf.FORM.block,
-        => blk: {
+        => {
             _ = try leb.readULEB128(u64, reader);
-            break :blk creader.bytes_read;
+            return creader.bytes_read;
         },
 
-        dwarf.FORM.sdata => blk: {
+        dwarf.FORM.sdata => {
             _ = try leb.readILEB128(i64, reader);
-            break :blk creader.bytes_read;
+            return creader.bytes_read;
         },
 
-        dwarf.FORM.string => blk: {
+        dwarf.FORM.string => {
             var count: usize = 0;
             while (true) : (count += 1) {
                 const byte = try reader.readByte();
                 if (byte == 0x0) break;
             }
-            break :blk count;
+            return count;
         },
 
         else => {
             log.err("unhandled DW_FORM_* value with identifier {x}", .{form});
             return error.UnhandledDwFormValue;
         },
-    };
+    }
 }
 
 fn findAbbrevEntrySize(self: DwarfInfo, da_off: usize, da_len: usize, di_off: usize, cuh: CompileUnit.Header) !usize {
@@ -455,7 +455,7 @@ fn findAbbrevEntrySize(self: DwarfInfo, da_off: usize, da_len: usize, di_off: us
     const reader = creader.reader();
 
     const tag = try leb.readULEB128(u64, reader);
-    _ = switch (tag) {
+    switch (tag) {
         std.dwarf.TAG.const_type,
         std.dwarf.TAG.packed_type,
         std.dwarf.TAG.pointer_type,
@@ -464,9 +464,11 @@ fn findAbbrevEntrySize(self: DwarfInfo, da_off: usize, da_len: usize, di_off: us
         std.dwarf.TAG.rvalue_reference_type,
         std.dwarf.TAG.shared_type,
         std.dwarf.TAG.volatile_type,
-        => if (creader.bytes_read != da_len) try reader.readByte(),
-        else => try reader.readByte(),
-    };
+        => if (creader.bytes_read != da_len) {
+            _ = try reader.readByte();
+        },
+        else => _ = try reader.readByte(),
+    }
 
     var len: usize = 0;
     while (creader.bytes_read < debug_abbrev.len) {